Created new public methods for the Form class:
Form::insertBefore() and Form::insertAfter():

The functionality will be useful for adding things like error messages to an existing form.

The class will also now throw an exception if you try to add a component to the form with
an index that is already registered with the form.

To mitigate, use Form::hasComponent().