Introduced Namespaces
Minor refactoring of the file structure. Inputs are now a sub package of
Components.

A few class names have changed  as a result of the namespaces, but this
should not affect any of the core classes as these were only abstract
classes.

If you have created your own FormComponents, they may need to be
revisited after this update.